Support StandardValueSet fields update
Up until now, we treated all CustomObject fields as CustomField
However, if it is not a custom field (e.g. LeadSource),
We need to handle it accordingly when updating.
This commit handles StandardValueSet (standard picklists), but
we might find out there are other standard fields that require
different handling.

Solves JIRA Issue: SALTO-81