Cli - On the way to Awesome (#260)
*** Help text after errors**

Added the help text of the failed command after the specific error.
How? Disabled the default help by yargs and added help output after errors (and the way we handle error/s generally).

** Also removed unwanted texts from no params text

*** Fix positional-arguments help message issue (was shown with --)**

How? Changed the way we color things which was causing the bug.

*** Change commands order in help message**

How? Added orderRank to commands and sort by it

* Some styling tweaks (coloring/bold etc.)
* args naming and order changes by convention
* Remove duplication of error printing when there's an exception (and clean it a bit)
* Added "Usage: " prefix to the help command
* Changed "Positionals" to "Arguments"