Handle compound fields (#317)
This PR adds support for compound fields. The change includes:

From now on compound fields will be discovered and shown in field_types.bp
Compound fields' nested fields will not be shown in other bp files.
Added support for compound fields for data migration (export will now show the nested fields of the compound fields and not the JSON compound fields). These are the same fields that the API expects during import.