fixed: CI build breaks on file renames
The CI build process runs with cached copy of the dist directory of each package.
This broke when files were renamed, because the old files were left in the dist cache.

Fixed by not including the `packages/*/dist` directory in the build cache, effectively
disabling the tsc cache mechanism in CI.