CI: proper coverage check
fixed - running yarn test in the repo root was not checking coverage.

since this was going on for a long time, coverage has deteriorated considerably for most packages, below the thresholds specified in each package's jest.config.js. fixed by setting the thresholds to their current values. this should be increased over time