Fix dummy adapter returning fragments instead of elements
Fixing a bug where the dummy adapter would return parse results as elements.
Parse results are fragments with incomplete types and cannot be treated as elements.
This quick "fix" is to resolve the types on the fragments but this is not a proper
fix because the adapter should not be using the parser in the first place. if it needs
to load elements it should be using an element source...