hotfix for missing static file when generating from a current env that does not have the files and cache invalidation takes place (#2471)
When a cache invalidation takes place, the multi env source will attempt to recover the files while using the wrong static file source. If the files are not there - missing file error will be created.

The following PR is a hotfix for this issue which does not address the real root cause (wrong multi env source is used during cache invalidation in the workspace). The root cause will be fixed in another PR later on today. (And this PR will be reverted)