Handle conflicts when both sides add an element
Before this change, if both sides (nacl and account) added an element, we would not consider it a conflict
even if the values inside were completely different.
this changes the behavior such that each difference between the added values would be considered a conflict
with the exceptions of values that we added by the account - those will be applied anyway.
This is not ideal, but it could help us avoid false conflicts due to hidden values that come from the account