Fix nacl update file picking logic
When updating a nacl file, the changes need to be routed to a specific file name.
In order to support filesystems where the file name is case insensitive, we group changes
in a case insensitive way.
However, before this change, that logic did not consider pre-existing files.
This made it possible for files to be created with different cases if the two cases arrived
in different calls to updateNaclFiles.