Fix unneeded cache invalidation in multi-env source
Before this change, if there was no semantic change in a nacl file source, the other caches
(multi env and workspace) would not be updated, which meant the hash would become out of sync
leading to incorrect cache invalidation the next time a semantic change is detected.

The fix here keeps the hash as a non-semantic meaning (unfortuntaly) but hopefully ensures
that it is updated even when there are no semantic changes.
So while there are still many different hashes that can all apply to the same semantic value
which is not ideal, hopefully now the nacl source and multi env source would both be updated
so they do not lose sync of which hash is the current hash