Skip building a source map when parsing NaCl unless it is really needed
Building a source map during NaCl parse makes parsing significantly slower.
It also greatly increases memory consumption.
It is also not actually needed in most flows and it is actually usually discarded.

Exposing the internal option to skip generating a source map and using it to improve overall perforamnce
when loading a workspace without a cache