SALTO-2234: Salesforce - Handle LIMIT_EXCEEDED on retrieve (try 2)  (#3014)
If we get an "LIMIT_EXCEEDED" error on retrieve, try requesting less items in each request. If reducing the count worked, return a suggestion for a new configuration for max items in retrieve request.

We keep minimum allowed items 500, although some items such as LWCs can be quite large (>50MB each) and the limit is around 600MB per response. This means that we might retry some items on each request until they are contained in a small enough request, but we will fetch all the small instances quickly.

Items which fail due to size once we tried them on a separate request, will be added to the skip list.

Currently the implementation is very simple, and will reduce the overall limit on items for ever even if we have a very small number of large items. In the future, we probably want to make this much more fine grained, and allow chunk sizes per type (or even mark specific instances to be retrieved individually), this way we won't slow down everything due to a few large instances.