Fix possible infinite loop on restore values
Resolved reference values can be fully resolved elements.
In such a case, using a naive compare would lead us to traversing all the references
in those elements, and in the case of a reference cycle, this traversal can be infinite.

Added custom logic for comparing elements in this scenario (when we know they are the value of a reference)