SALTO-2037: Remove XML Attribute prefix only for fields we will convert back when deploying (#2972)
For Lightning Web Components, we convert known attributes to look like regular fields, and convert them
back to attributes when generating the XML for deployment. We can't do this for arbitrary attributes which
are not marked as "IS_ATTRIBUTE" on the type, since we won't be able to covert them back.