SALTO-2339 - Fetch From Workspace fromState static files fix (#3058)
Before this Fetch From Workspace, when being used with the fromState flag, did not fetch the actual content of Static Files that are part of the changes fetched. This is cause the content is not available in the state.

This uses the sourceWorkspace’s staticFileSource as the source and to make sure the hashes match (+ remove the change and add an error if it doesn’t).
This also adds a new function to workspace, getStaticFile that is being used here.