SALTO-4120 - Change getFromPathIndex to not return multiple hints for elemIDs that do not exist in the pathIndex (#4348)
We have an issue when using splitElementByPath and using a pathIndex that does not have information for all the elemIDs that are being split. For example splitting an Object type and one if the fields does not exist in the pathIndex.
This is caused by the fact that getFromPathIndex returns the hints of the "closest" parent, even if that parent has multiple hints.
This changes the logic so that for elemIDs that exist in the pathIndex, we will still return all the relevant hints, but for those who do not, we will still use the closest parent's hints, but only one of them and not all the hints cause returning multiple will eventually cause merge errors for this element since the part that does not exist in the pathIndex will appear in multiple fragments.