SALTO-3817: Add postAction.showOnFailure flag
The post-deploy action messages for the salesforce adapter should be shown if
the change failed but the other ones it should not.

To deal with this I've added a flag named `showOnFailure` to the `DeployAction` type (by
creating a subtype called `PostDeployAction`).

When showing the post deploy messages I filter out ChangeError instances of failed
changes where the `showOnFailure` flag is true.

Once this is merged the next step is to handle this flag in the SASS part.