Change regex for identifying dynamic content and references
The current regex allows for some pretty whacky stuff - for example, having `{{ blip bloop blep dc.content_item }}` will catch `dc.content_item` as a dynamic content placeholder, even though Zendesk wouldn't allow it.

In this PR, i've changed the regex so that it forces the content to be in brackets, and furthermore forces the brackets to be "close" to the placeholder (so if you have content between the brackets and the placeholder it won't catch it anymroe).

This issue was caught when a customer tried to note his email, which had the string `dc.stuttgart` in it, and it got replaced by a dynamic content ref.