SALTO-5773: Bumping fast-xml-parser
This is a major which required a bunch of changes. The reason for the bump is the fix for https://github.com/NaturalIntelligence/fast-xml-parser/issues/350 which solves the bug, but there are also a lot of additional bugfixes.
Also had to copy some code from the SF E2E utils to the UT utils to get the linter to stop complaining about the TS scope.