SALTO-6699: Cache whether a remote map is empty
Checking if a remote map is empty requires creating a DB iterator with a prefix.
If we know that a namespace is empty, we can avoid doing that and hopefully reduce the amount of DB iterators we create.
This in turn will hopefully reduce the chances that we trigger the crazy memory consumption issue in rocksdb