SALTO-6903: Avoid creating a unified source map for many files
When finding locations for changes, try to avoid creating a single source map.
We suspect this single source map consumes a lot of memory.

Instead, try to group changes by which files that could be located in, and for each group
create a source map just from the relevant file(s).

This is all behind a flag: SALTO_USE_SPLIT_SOURCE_MAP_IN_UPDATE