SALTO-7047: Improve performance of workspace update
The override mechanism is replaced by fixing the original element source to return the correct result
The issue that the override mechanism was there to solve was that if a field was removed from NaCl, we should remove the hidden part of the field as well.

This didn't work in the original code because the state element source would try to merge it with the element from the workspace cache
but the element in the workspace cache was not updated yet, so it still had that field.

Changing the element source to use the element from the nacl cache instead (which is updated before) should solve the issue