SALTO-7188: Setting flow author info from selected version
Currently the author info is derived from the flow object which is not necessarily what users expect. Creation info is for when the entire flow was created and change info is for the last time a version was created or activated.
Now, the author information for the selected version (latest or active, depending on user preference) is used. Creation info is from when the version was created and change info is from one second later or when the version was activated (this is SF behavior).
Also, no longer hiding the FlowDefinition type because that's covered by the hide types folder filter.